pyproject.toml
[project]
name = "unit-standardizer"
version = "0.1.0"
description = "entity traversal based on benchling schema"
readme = "README.md"
requires-python = ">=3.10"
packages=["unit_standardizer"]
dependencies = [
    "antlr4-python3-runtime>=4.13.2",
]

[project.scripts]
unit-standardizer = "unit_standardizer.run:main"

[dependency-groups]
dev = [
    "ruff>=0.12.10",
    "pytest-cov>=6.2.1",
    "tox>=4.28.4",
    "tox-uv>=1.28.0",
    "mypy>=1.17.1",
    "pre-commit>=4.3.0",
    "notebook>=7.4.5",
    "python-dotenv>=1.1.1",
    "pytest-mock>=3.15.0",
    "ipywidgets>=8.1.7",
    "antlr4-tools>=0.2.2",
]

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"
